Add option to skip models that report being Succeeded when starting builds.
CHANGELOG: `genome model build start`: new option --skip-succeeded. If
specified the command will skip starting a build for any model with a
status of "Succeeded", just as it does by default for models with a build
that is "Scheduled" or "Running".

This is intended to streamline the use-case of "start new builds for any
models in a model-group that haven't yet succeeded".