Updates to the GenePrediction::Bacterial build process.
This pipeline no longer skips PAP, so it now fully emulates the non-pipelined
version.  Obvious future development would be to have it work more like the
ProteinAnnotation pipeline instead of being a thin wrapper around the old
scripts--these changes are mostly narrowly targeted at being able to
systematically run tests for changes to the scripts (and maybe lead to the
eventual migration to this model-build process).