Change VcfToVariantMatrix to write to many tmpfiles
previously, this command would write partial output, compute more, read
in the previous output, append to it in memory, and write it all back
out again. That leads to O(n^2) complexity in the number of data chunks.
If we instead write out each chunk to a temp file, then merge at the
end, we get O(n).