remove MAX_VOLUMES limitation
If we don't want to allocate to volumes we should just manage the
volumes better rather than have code accounting for that. Additionally,
the idea is complicated by the lockless algorithm. We would prefer thing
not compete for the same volumes to improve performance though with
lockless algorithm it is only really relevant as a volume reaches
soft_limit_kb.