Added wait_for_lsf_job_ids to Genome::Sys
This method takes a list of LSF job ids and waits for all of them to
finish before returning. Return value is a hash of job_id => status
pairs. Also added test to cover this addition, though it will not be run
as part of the automated test suite because it requires LSF jobs to be
scheduled which can hang for a number of reasons.