Refactored logic into helper function Change-Id: I6bc17bc75d298c7aefa9e0ad94808c244480eefa
removed unused lib and whitespace
Change-Id: Iec797bf4e4487dc03a95ec70c5e8442a5279aa2b

Add --bowtie-version specification to chimerascan params
You can now do something like:
  fusion_detection_strategy:
    'chimerascan -v 0.4.5 [-p 2 --bowtie-version=0.12.7]'
and only '-p 2' will be passed on to chimerascan and we take care
of telling chimerascan what version of bowtie to use.

You must supply a bowtie-version, and it currently must be of
major version 0.

Change-Id: I179271018de084d8a8f9caa713094a3347a39678