relationship therapy - MergeAndFixVcfs
in gmt relationship merge-and-fix-vcfs

improve performance by keeping an iterator into the current chromosome
for the denovo file rather than looping over all denovo positions in
said chromosome for each position in the standard file.

before switching to a new chromosome, dump all remaining (unprocessed)
denovo sites to the output. previously, these were dumped at the very
end of the file, so if the denovo file contained sites on chr1 with pos
greater than any site in the standard file on chr1, they would appear
at the end of the file rather than prior to chr2. this change makes the
output much closer to sorted, but...

joinx sort the final output file, since the chromosome order of the
input files depends on the reference sequence in use. the ticket that
spawned this work was using hg18, so X,Y are found between 9,10, which
may cause problems down the road.

this (hopefully) fixes rt#86339