remove unused properties
These properties were not defined as inputs or params which means they
would not be saved to the database. It was brought to our attention that
several classes mistakenly forgot to set properties as inputs or params.
To make it explicit properties should be labeled as transient. Since
they also appeared to be unused I removed them instead.

Change-Id: Idbf1f887744ec3036a8989f0514bf43498a0d5ac