explicitly label properties as transient
These properties were not defined as inputs or params which means they
would not be saved to the database. It was brought to our attention that
several classes mistakenly forgot to set properties as inputs or params.
These looked like they were not intended to be saved to the database so
they were set to transient.

Change-Id: I62ee66c6c86b05385c3625d65e5533f83a9f1f45