remove unused property
These properties were not defined as inputs or params which means they
would not be saved to the database. It was brought to our attention that
several classes mistakenly forgot to set properties as inputs or params.
This property appeared not to be used and was instead "overridden" by a
sub.

Change-Id: I6cd0afd3f3d170d0db21edb4e8220c68f9c6ad15