try to optimize SR get queries
Several SR->get() queries were using lots of resources in oracle,
causing misc. havoc throughout the center.  This is an attempt to reduce
that resource usage.

DETAILS

The queries used to look up SRs involved many joins to the
software_result table, and many of those involved params which were
non-specific to the SR being looked for (e.g. aligner_name,
aligner_version).  To reduce the number of joins and merges we've done
an initial query on only the inputs, then searched through those results
by hand to find the full result.