add input blacklist and params whitelist to SR::_faster_get
This will make us fall back on old behavior when a software result has
only inputs of the specified name.  This is desirable, because the new
get is slow in cases where the inputs correspond to many software
results.

Additionally, add params whitelist so we can search on highly selective
params by default.