Improved clinseq parallelization.
The ClinSeq (to be renamed) pipeline now launches SummarizeSvs and
SummarizeCnvs in parallel, as independent workflow operations, and
uses a dynamically generated workflow internally instead of the
_execute_build API.

The Main step now has two outputs, one for the input directory of
each of the subsequent steps.

The Main step now does more on a "dry run" execution, testing
the lighter-weight parts of main inside of ClinSeq.t.  The "dry run"
version of the pipeline still skips the rest of the worflow so that
those steps do not need to have special "do nothing" logic.

This commit depends on an update to Workflow which ensures that
the Command adaptor (Workflow::OperationType::Command) treats both
params and inputs as operation inputs.