optimize for models with many builds
The old method would load all builds or more specifically all the build
events in order to pop the latest one off and return it. This now uses
an event iterator for a significant performance improvement. A build
iterator was tried but did not work due to the status and date_scheduled
being on the_master_event.