Cleanup of Genome::Model::Command::Define::BaseMinimal
The way that this class copied model property data led to shared
references, and too many attributes being passed through.  This is now
fixed.

In addition, the base definer now lets the model validate its own
processing profile and inputs instead of doing it in the command.

The _validate_inputs, validate_processing_profile, and
compare_pp_and_model_type methods are now removed from this class
and also the old ::Helper subclass from which it was pulled-up.