cleanup Genome::Model
The updated Genome::Model class now has doc lines for everything.

The subject_class_name property is now an indirect accessor instead of
r/w.  The column in the db (non-nullable) is now bound to
_subject_class_name until all old snapshots are removed, and it can
be dropped.

Properties which are deprecated but still needed by things which
are not using ModelDeprecated as a base class (tweeners) are
in their own section.

Pipelines such as ClinSeq can now abandon ModelDeprecated completely,
and hopefully somatic variation will follow soon.

Notes are present throughout the module as to the status of legacy
parts.