Fixed data loss issue when creating allocations
Fixed an issue where data loss could occur when creating allocations. If
the allocation system attempted to create an allocation that had an
absolute path that already existed on the file system (and that path did
not already have an allocation), that path would be deleted when the
allocation system created and soon after deleted a "candidate"
allocation for that path. Creating an allocation now checks for this
condition.