Fix 'genome model define reference-alignment' default
A new feature of UR is the ablility to add objects as defaults, one side
effect of that is that UR actually cares what you put as defaults and
uses those defaults to look up the value.  The old default for
reference-sequence was just the name of the reference sequence... Then I
changed it to be {name => 'some_name'}.  All seemed well until the
Define-docs.t test fails and you find out that the hash memory address
was being displayed and therefore hashes are disallowed as default
values.  In the end I just replaced the name with the id.  It is less
understandable to an end-user but I don't know what else to do.