Read generation for PerLaneTophat aligner index
Genome::InstrumentData::AlignmentResult::PerLaneTophat requires several
'fake' reads during index creation. We previously had 3 hard-coded
sequences taken from the human reference, but this caused the command
used for index creation to give a non-zero exit status when run against
a reference that did not have these sequences. Instead of trying to
verify that the index was created despite the exit status, we now
generate 50 additional fake reads that are appropriate for the provided
reference and annotation.