restore default_values
This tool is using defaults to find compatible models or creating new
models if needed (which would then validate their inputs as well).  So I
think it may be an exception to the "no defaults" rule.

That commit did not convert several defaults which possibly should have
been converted (if this is ever unreverted).

This reverts commit 935b04397f9b925fac4b80aacbdcbcf79d70dea8.