re-enable test
Upstream builds finished building so re-enabled test.  Add an additional
test because it now get's the model's latest build instead of a specific
build.  Updated expected output which was mostly due to changes in build
IDs and paths but there were also a few changes in output that looked
minor.