hide color when STDOUT and STDERR are not TTYs
When piping or `watch`ing commands that use this the color escape codes
would be visible.  I don't know if this is the best way to do this but
it seems to work.  Since we only have the string we are assuming we
should disable color if either STDOUT or STDERR are not TTYs.