Genome::Command::Tester from the old clinseq tester module.
This module exports the run_and_diff() function.

Now updated to get test the class name matching the test name by
default.

In this case, or when a class name is specified instead of a full
command, it runs the help_synopsis(), replacing /tmp/output_dir
with the desired output directory.