Switch to use the Dwrac data source
This also nukes the union that pulls in Sanger rows.  It was technically
a problem because of the join on misc_attribute in the main Genome database
with joins to Oltp and Dwrac on either side.

The conclusion was that this class is only used by the data importer that
migrates data from the LIMS database to Genome's, and Sanger data is
handled via a completely different process - getting rows from the
sequence_read table, making allocations for them and inserting them
as imported instrument data.