Remove UR_DBI_NO_COMMIT from TestHelpers.pm
since all .pm files are loaded by commands as they search for
sub-commands, we cannot put this in TestHelpers.pm.  It would be a nice
thing to have there since it would stop people from putting test
entities into the database by accident, but until we can safely put it
there it should be removed.