mark many GENOME_* variables as required
If we defined a value in Genome::Site::TGI and a default_value was not
supplied then these are now "required" configuration.  A few exceptions
were:

GENOME_DEV_MODE
GENOME_STATSD_HOST
GENOME_STATSD_PORT

We can decide later whether not having these set should be a
compile-time or run-time error.  For now we have to be careful about
failing when they are not set since the code that uses them is probably
not very defensive about it.