Auto-increment names for LaneQC models.
We'll now auto-increment names for LaneQC models like we do other
models. This is expected to have a few side effects, but we'll
address them as they come up. We have a situation where a piece
of instrument data could have multiple valid LaneQC models from
being used in multiple analysis projects which is something that
hasn't occured previously.