bugfixes in the test
The gatk_results need to be uniquified between the different subtests.
This is achieved by adding the know_sites build to the retrieval method.

For the gatk and clipoverlap combined test the retrieval of the clip
overlap result should use the gatk results since that is the previous
step (not the merge).