remove ability to try to lock the same resource twice
I searched for "wait_on_self" and only found it set in its own test.
With wait_on_self defaulting to false if something tried to lock the
same resource twice it would unlock its current lock and then acquire a
new lock.  This should almost certainly be a fatal error because it
probably means something else forgot to unlock its lock or if this is a
means of "sharing" a lock it means something else could acquire the lock
out from under this process.

CHANGELOG: It is no longer possible to lock the same resource twice (in
the same process).