Change inputs to AnnotateWithReadcounts
We cannot rely on UR to keep is_many properties in any particular order,
so we have to combine these two fields.  While we're at it, since
sample_names aren't in fact useful except to get at the column index in
the vcf-header, lets go ahead and just require the index.