Use awk instead of VcfToBed to create a region file.
VcfToBed is not the tool we want for this. We don't actually want a bed
file, we want a one-based region file. Moreover we do not actually care
about sample information and their genotypes. We just want the positions
where ANY variant was called. Awk is the easiest way to do this.