Correct bug in File::Vcf::Entry. Change behavior when no sample present.
File::Vcf::Entry contained a bug where the bounds of the sample index
was too permissive by 1.

We also now return undef instead of confessing when we have no GT
information for a sample - the sample might just be '.'.

Corrected an oversight in the unit test - there were not enough samples
in the header for the "get genotype for sample" subtest data. And we had
to correct the behavior assumptions when dying for the above.