change query_tcga_barcode timeout, sleep, and attempts
Previous, default timeout was 180s, sleep was 60s, and attempts was 5.
This mean query_tcga_barcode would hang for 20 minutes is the TCGA UUID
Browser was down.  Instrumentation so far shows that so far it has always
either succeeded on the first attempt or failed and that the upper bound
on average successful response times was 700 ms.

Now the default timeout is 5s (much greater than 700ms), sleep is 10s,
and attempts is 8 which means query_tcga_barcode will hang for 2 minutes
if the TCGA UUID Browser is down.

This will prevent tests from failing and should, presumably, not cause
an increase in failure rates (due to the first attempt succeeds or it
fails).  Future instrumentation values will allow us to verify.