Use the build directory to decide which result is the correct one.
If the indel_detection_strategy contains an unfiltered detector that
also detects snvs, the previous logic would get confused about which
result is the correct one.  The final result symlinked in the variants
directory knows the one true result.

(In SomVal a different label is set on the SR::User--this would be a
better solution, but it isn't backwards-compatible.)