Remove available_fields where possible
available_fields is defined in the base interpreter class to get
its values from the field_descriptions hash.
If the field_descriptions hash contained references to $self this
would fail because during validation we were calling available_fields
as a class method.
Now that available_fields is exclusively called as an object method
we can use the base class version of available_fields whenever
field_descriptions is available.