CHANGELOG VariantReporting resources changed to translations only
In an attempt to simplify VariantReporting, the "resource file"
now only provides "translations".  Any resources previously in the
resource file should now appear in the plan file with the value being
an appropriate variable name.  This name will appear as a key in the
resource file, and the actual value will be translated