Remove UR::Context::Transaction from Command::DelegatesToResult
The use of the transaction exposed cases where we were incompletely
constructing objects.  These objects are not DB backed and their
references fall out of scope, but the transaction still knows about them
and when we commit it determines that they are in an un-savable state.

All of those cases need to be addressed before we can use a transaction
here.