Don't cache available_fields_dict
This was causing some subtle problems with the plan validation tests.
The issue is that UR::Object::create_entity (which is called by
->create as well as ->__define__) calls __errors__ on the object.  That
would cause the available_fields_dict to be cached BEFORE the
interpreters and filters were added to the Reporter object.

Since I don't know any reason (other than efficiency) for memoizing this
method, I think its best to just remove it.