increase threshold
This is mostly arbitrary but we want to disable volume that are unlikely
to be able to fit anymore data on them.  The more we scale this
threshold the more space we (temporarily) waste but the lesser we scale
this threshold the higher the chance of keeping unusable volumes in the
candidate volume pool.