Add a bam readcount parser that can encode and decode bam-readcount lines in vcf
Lines from bam-readcount are stored in a sample format field as json.
The json contains a hash with two types of keys:
-alternate alleles.  The values of these entries point to which offset key
should be checked to find the bam-readcount line for that allele.
-offsets.  The key is the offset from the vcf position on this line (since
bam readcount interprets deletion start positions differently, they will be 1; all
other types of normalized vcf alleles will be 0).  The value is the bam-readcount line.

The json is made vcf sample field safe using the encode function