New module to compare contents of two directories
This new module Genome::Utility::File::DirCompare was adapted from File::DirCompare
(http://search.cpan.org/~gavinc/File-DirCompare-0.7/DirCompare.pm)
If two paths are both symlinks, this module will resolve the symlinks and do
a file comparison on the targets. The previous behavior in File::DirCompare
for two symlinks was to count them as different if they were pointing to
different targets. With this change they will only be counted as different if
the contents of the targets are different.