Some wider searching for a matching reference.
Since combining two is the most common case, we're willing to search
a little more deeply for a common reference.  The number of likely
candidates is still small.

The common scenario here is that the vendor supplies a BED file using a
reference with "chr"s present, but our combined BEDs for analysis do not
have the "chr"s.  Thus a conversion to the "chr"-free form is required
before combination.