protect against corrupted Workflow::Models
If _find_first_nondone_step_impl crashes it can leave the related
Workflow::Models in a corrupt state that ends up having "invalid data
for save".  The (side-effect) changes are probably a bug and not
relevant to the task at hand so it's safest just to rollback if
_find_first_nondone_step_impl crashes.

The "invalid data for save" was already being rolled back later when the
transaction fails to commit but by pre-emptively doing it here we can
save a ton of time due to the slow performance of the "invalid data for
save" code.