Properly scope the saved STDOUT and STDERR handles
This should fix the two tests hanging on Jenkins:
    InstrumentData/Command/Import/Basic_sra.t
    InstrumentData/Command/Import/WorkFlow/SraToBam.t

The problem is that Genome::InstrumentData::Command::Import::WorkFlow::SraToBam::do_shellcmd_with_stdout()
is reading from a file of saved STDERR of a child process run with
Genome::Sys::shellcmd() and printing it to its own STDERR.  However, the
current process' STDERR is also printing to that same file.  The tests hang
because do_shellcmd_with_stdout() never gets to the end of the file (it's
filling it as it's reading).

I'm not sure what the _actual_ bug is, but this seems to fix it, while
removing global filehandle variables.