Fix the Sys-shellcmd.t 'Run echo with stdout redirected' test
We only need to print a newline if STDOUT was _not_ redirected.  This is to
compensate in case the child process didn't print a newline after it's last
print, which would screw up TAP in the parent process (which prints to
STDOUT).  If STDOUT is redirected, then whatever the child process prints
there won't have any effect on the parent process' TAP output.

Previous shellcmd() code waited until the child program was done, then
printed a newline to STDOUT while it was still possibly redirected.