Convert Genome::Config::AnalysisMenu::Command::Item::Create to do the right thing and add a meaningful test
AnalysisMenu::Item->create now requires a pull request for the menu item repo
to be merged and the 'Update Analysis Menu Items' jenkins job to have run
before this command can be used to create a menu item.  If the file being used
for menu item creation has not be added this way, this command will fail.