override default timeout when creating an ephemeral database
We've had these errors every once in awhile,

> Got error response 500: SSL read timeout:  at
> /usr/bin/test-db-database-create line 15

Previously we added the `--timeout` option but we didn't think it should
be necessary.  Now I am thinking it may be that the template locks while
creating the new database so contention is what is causing the timeouts.
For now I am just increasing the timeout as a workaround.  We'll need to
see if this is actually what's going on.