Allow for multiple WorkflowBuilder backends
The interface for the 'execute' method changes here as well.  The old
interface was to accept any number of named arguments, but this makes it
so that the interface cannot be extended in the future.  If one were to
ever want to add an optional argument to execute, it would collide with
the workflow inputs.