add MigrateLockProxy
A MigrateLockProxy object can be used to migrate a lock from one
resource/scope to another.  This will aid in fixing issues where the
chosen resource turns out to be a poor choice, e.g. it's length exceeds
that supported by a backend, and with migrating away from the `tgisan`
and `unknown` scopes.