Add ability to set exit code
This command needs to sometimes exit 1 (fail) and sometimes exit 0
(succeed) regardless of it successfully doing its work.

It will be used similarly to a 'shortcut' step in a workflow, and should
always exit 1 when used in that way.

It will also be used normally in a workflow, and should always exit 0
when used in that way.