avoid tight coupling to last error_message
These tests fails for genome/UR#75 because a later error message is sent
that differs from the one these are testing.  The ones they are testing
are still sent but are not the last `error_message`.  I've changed the
tests to try to avoid tight coupling with the last `error_message`.