change `env` prefixes to XGENOME_
In order to break ties with the legacy config system I am changing the
prefix for environment variables from `GENOME_` to `XGENOME_`.  This may
mean that some scripts start behaving differently due to configuration
not taking effect but it will allow us to immediately avoid the
confusion of whether a value was set by our default or by the user and
will allow us to change defaults without having to tell users to start
new shells, etc.