Make Builds use a Genome::Process when run
This should be invisible to users, except for a few debug lines
that get emitted when you run 'genome model build start'.

* Logs still go in the usual place (build's data directory).
* Workflow job has same lsf_project name.
* Workflow has the same name (meaning genome model build view works
as expected).
* Emails are still sent when the build starts, fails or succeeds.

Command 'genome model services build run' is left only because
job_dispatch=inline is used in various tests.  Since PTero will not
support this, we will eventually have to update all places that utilize
that feature.