Get operationtype attributes from the xml
These attributes:
    - lsf_project
    - lsf_queue
    - lsf_resource

were only being found (if they existed) on the Command class itself.
Therefore when you read in a workflow, and it had specified any of these
values, they were ignored.

The hash EXPECTED_ATTRIBUTES was changed to a method so we could move the
functionality for finding these attributes in the xml up into the base
class.