Set workflow_builder_backend outside of begin
Setting 'workflow_builder_backend' inside the BEGIN block prevents the
new setting from taking effect outside of the BEGIN block.  Before this
change, the 'workflow_builder_backend' reverted back to 'workflow',
which is clearly wrong for these Ptero integration tests.

This bug remained uncaught for nearly two months because of another bug.
The configuration of the Genome Ptero Integration had a typo which
caused the tests to only run against a particular commit of the genome
repository, a commit which was from April 2015.