Delete Genome::Processes tied to allocations
Before this change, there would be a constraint violation since the Process's
refers to the to-be-deleted allocation, and the allocation wouldn't be
deleted.

Now, the constriaint violation still prints out, but the allocation is deleted.
Fixing this probably requires either calling sync_database() after deleting
the Process object, or asking the allocation deleter to cascade its deletion
to a related Process.  Neither option feels right...