Don't require builds to be from the same model to diff
We have a use case for CLE regression testing where
we are making new models in a testdb and want to diff
the resulting builds with builds from old models.  They
should be equivalent, but not the same model.
I don't see any reason why this check adds to the
integrity of the diff