New interface and implementation for get_library_name_labels
The previous implementation assigns a label for each library
that the sample has in the database.  We actually only want labels
for the libraries that are relevant to a set of builds, which may
be a subset of all possible libraries for a sample.
This will change the number of indices and may change the
blessed Trio process.