GMT Analysis Cov Bam RC
Revert checking to see if bam read count version supports
insertion_centric in the gmt analysis cov bam. This will get passed to
the gmt bam read count and ultimately fail. This is required at this time to catch the
instances of bam read count v0.4 being used in this process.

This reverts commit 551627d2f81c17d56ecd77981fb09dcb30cc8c60.