Add converter to remove extra contigs
Implementing this as a whole-file converter as dropping lines seemed
kludgy in the current design. It also was leading me towards a global
for the chromosome set needed to check valid chromosomes and that seemed
bad. The use of resource is a bit hacky, but allows us to do the right
thing.