Allow evaluation to be re-parsed
If the evaluation has already been run and we just want to
get the stats, we don't need to re-run everything, which takes
a while.  Detect this situation if the compare.txt file exists with
size.
The disadvantage here is that if you actually do want to rerun (e.g.
something has changed), it may not do what you expect (it does print
a debug message that results are just being parsed).
Previously, it would crash when trying to create a symlink that
already existed.