Make the volume lister work
Disk Groups validate that their disk_group_name belongs to a short list
and dies otherwise to ensure Genome stuff only messes with analysis
allocations.  For a basic lister, we don't want want this safety net.

Getting the free space of an unmounted volume also throws an exception
if the volume is unmounted to avoid mounting tons of volumes.  For this
simple lister, we just print out "<unmounted>" for such volumes instead of
dying.