add method bsub_and_wait_for_completion() to Genome::Sys
Used to submit many jobs to LSF and wait on their completion in an efficient
way.

Existing code will submit jobs and then wait on all of them with
wait_for_lsf_jobs() which busy-waits by calling 'bjobs' over and over.
There's an issue with this pattern where the continual running of 'bjobs'
can cause network sockets on the LSF server to pile up in TIME_WAIT

This method avoids the problem by setting a post-exec command on each
submitted job that notifies the submitter when the job is complete.  It
can then do (optimistically) one pass of 'bjobs' to get the completion status.