Stop calling delete() when creating a Model fails
Instead, do the creation inside a software transaction we can roll back if
there's a problem.

This fixes the potential problem of someone trying to create a new model with
the ID of an existing model, failing, and then having the delete() cascade to
delete all the builds of the previously existing model.