Remove the construct_lsf_param_string functionality
This function worked as long as no values in the param structure needed
to be translated to construct the lsf param string.

The units of the value associated with the -c option should be minutes
in the lsf param string and seconds in the lsf param structure, which
means a translation of the value would be required to construct the lsf
param string.

Instead of implementing this translation for the -c option, I'm removing
the construct_lsf_param_string function, since nobody uses it anyway.