Run the HaplotypeCaller serially on each contig in a bucket.
It has been decided that it is preferable to produce per-chromosome
GVCFs rather than have collections of chromosomes bundled together.
In order to still have consistently-sized jobs (and avoid splitting
hundreds of ways, which will pound the source BAM file), run the
HaplotypeCaller on each entry in the bucket in turn.