Get the Sample/Command/Copy.t test working
In the original code, Genome::Sample inherits from Genome::Subject and
Genome::Utility::ObjectWithLockedConstruction, both of which implement
create().  Since Subject appears in the 'is' list first, that's the create()
that gets called within the call to copy().

That's a bug, because then ObjectWithLockedConstruction doesn't influence
Genome::Sample at all.

This fixes the bug by wrapping G::Subject::create() inside G::Sample to do
the locking