Apply these roles to the concrete classes instead of the abstract parent
UR's roles now check for the existence of methods on parent classes'
namespaces in addition to the check on the consuming class's namespace.
This means you shouldn't apply a role with methods to both a parent and child
class.

In this case, the roles need to be applied to the child classes because the
parent class does not have a table_name, and the role intends to add
properties to the consuming classes that get saved to the DB.