Get the test working
Because the bug mentioned in f6ce90de3 is fixed, the test was dying in the
call to Genome::Model::Tools::DetectVariants2::Result::Manual->create() because
it detected the creation of a duplicate software result.  The cause was the
sample_id param.

In the original code, the above mentioned bug generated a new sample object
each time through the test_build() function, where in this fixed branch, the
Sample object was only created once, and returned over and over.

By using a new name each time, it generates a new Sample object, duplicate
software results are not created, and the test runs successfully.