Get the Logger test working
The test wasn't yet updated to override the correct function when the Logger
functionality was moved to a Role.  It was working when run from a real terminal,
but failed when run by the CI server, since it wasn't faked out to properly
believe it was on a color screen.