Don't die on empty input.
This step is run as a part of the LOH pipeline. Currently
if the variant file that is passed to this module is empty,
the workflow crashes with an error message(bam-readcount is invoked
with an empty file which causes it to die). Instead I think
we'd like this module to spit out a warning and not attempt to
filter if the input is empty.