Avoid calling delete() on a Set::Scalar::Null.
Although one can create a null Set::Scalar which can have delete()
called on it, the difference() method of Set::Scalar returns a special
subclass, Set::Scalar::Null, that doesn't have delete() defined.  This
avoids this possibility by skipping the deletion if the set is already
null.  (In that case, there's nothing to delete, anyway!)