Pass along objects as-is without further processing.
If someone already created the object for the get() call, we don't want
to end up creating a UR::Value of the memory address!  (This will
probably fail on non-UR objects unless they happen to support id(),
class(), and get().  But at least that'll be noisy instead of a quiet
failure.)