New WorkflowBuilder backend: "simple".
This `bsub`s a single job which then runs the workflow inline.  For
one-step workflows (like CwlPipeline or ReferenceSequence) this is
approximately the same as launching with Ptero, but without the
overhead!  For more complicated workflows (like SomaticValidation) this
would almost certainly be a poor choice.

This is pretty much equivalent to this combo of options from "Workflow":
   --server-dispatch lsf
   --job-dispatch inline