Don't use defined on an array.
It used to be a warning back when this was written. Now it's an error.
Either way, it doesn't make sense, so I borrowed the conditional from a
different line of the file to replace this one.

I make no guarantees of the ongoing usefulness of this tool, but it
compiles now :)