For now, don't try to `df` remote volumes.
For volumes that are mounted via a cache layer, df doesn't work
properly (among other non-compliant aspects of the filesystem).  In our
case we could possibly use SMB to connect to the remote server and find
out the canonical used and total KB, but for now let's trust our
database to be accurate.