Mitigate having multiple workflows per build.
The "right" answer here would be to link them to the corresponding
Genome::Process.  The issue right now is that the process is created
*after* the workflow inputs are assigned, so we have no way to
accurately determine the correct process for our current run.  There are
some strategies for trying to figure it out.  For example:
 - look up our current LSF Job ID and check the database for a matching
   process.
 - look for the one process that is currently marked as "running" and
   die if there's a second one.

For now this assumes only one process ever "succeeds" for a build and
the build view displays all the workflows in an arbitrary order.