Make the "prod-builder" run_as configurable.
Now that we're operating on multiple clusters we've run into a situation
where it'd be nice to distinguish between then with different "run_as"
indications.  This will let us change it when run on the new cluster
while leaving it the same on the old.  (Or someone enterprising could
set the appropriate environment variable to setup for the "other"
cluster for the one they're on!)