When call-caching is enabled, check the health of jobs.
Before we were letting jobs fail and manually running `script.submit`
for replacement jobs in the event of LSF failure due to the high cost of
starting a workflow from the beginning over such a trivial failure. With
call-caching that cost is reduced, so it's better to have quicker
notification of issues.